+++ debug/__init__.py
@@ -1,0 +1,3 @@
+from .player_log import PlayerDebugLogger, is_enabled
+
+__all__ = ["PlayerDebugLogger", "is_enabled"]